+++ Chalkable.Web/Models/DistrictViewData.cs
@@ -21,7 +21,7 @@ namespace Chalkable.Web.Models
             Id = district.Id;
             Name = district.Name;
             DbName = district.DbName;
-            SisSystemType = district.SisSystemType;
+            SisSystemType = (int)district.SisSystemType;
             SisUrl = district.SisUrl;
             SisUserName = district.SisUserName;
         }
